Add tests cases for coverage in expressiontypes.go FB-2041 Added tests for typeIsTimeQuantum and typeIsSet and made them pass. Added tests for DataTypeTuple cases in typesAreAssignmentCompatible. Timestamp conversion checking is handled before it gets to that point but I left those branches in as a backstop. Checking to see if DataType[String,ID]SetQuantum can be assigned to themselves doesn't appear to be reachable currently but left those branches in, because something may use them in future. Added one test to the DAX skip list since it's the IDSetQ version of a StringSetQ test that was already on there, changed skip list to refer to both tests by name instead of by number.

+++ sql3/planner/inbuiltfunctionsset.go
@@ -49,7 +49,7 @@ func (n *callPlanExpression) EvaluateSetContains(currentRow []interface{}) (inte
 				return nil, sql3.NewErrInternalf("unable to convert value")
 			}
 
-			return intSetContains(targetSet, int64(testValue)), nil
+			return intSetContains(targetSet, testValue), nil
 
 		default:
 			return nil, sql3.NewErrInternalf("unexpected data type '%T'", typ)
